requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
pillow==11.1.0
numpy==1.26.4
tensorflow-cpu==2.17.1
